@@ -4,8 +4,8 @@
     "type": "library",
     "keywords": ["phpunit","result", "printer", "unit", "test"],
     "require": {
-        "phpunit/phpunit": ">=4.0 || <6.*",
-        "php": ">=5.3.0"
+        "phpunit/phpunit": "6.*",
+        "php": "^7.0"
     },
     "autoload": {
         "psr-4": {
